refact: split overlay_backends into per-strategy package
The 721-line app/overlay_backends.py becomes a package with one module
per strategy (base, uno, custom, local) plus a small utils module for
shared helpers (is_custom_overlay, split_custom_oid, _mock_response).
__init__.py re-exports the public symbols so existing imports and the
@patch('app.overlay_backends.AppStorage.*') test fixtures keep working.

LocalOverlayBackend.is_visible() now reads from the OverlayStateStore
instead of a locally cached self._visible flag, making the store the
single source of truth for visibility across restarts.
